+++ example/lib/main.dart
@@ -1,6 +1,5 @@
 import 'package:curved_progress_bar/curved_progress_bar.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_screenutil/flutter_screenutil.dart';
 
 void main() {
   runApp(const DemoApp());
@@ -18,34 +17,34 @@ class DemoAppState extends State<DemoApp> {
   Widget build(BuildContext context) {
     return Scaffold(
       body: ListView(
-        children: [
+        children: const [
           SizedBox(
-            height: 200.sp,
-            width: 200.sp,
-            child: const CurvedCircularProgressIndicator(),
+            height: 200,
+            width: 200,
+            child: CurvedCircularProgressIndicator(),
           ),
           SizedBox(
-            height: 200.sp,
-            width: 200.sp,
+            height: 200,
+            width: 200,
             child: CurvedCircularProgressIndicator(
               value: 0.5,
-              animationDuration: const Duration(seconds: 3),
+              animationDuration: Duration(seconds: 3),
               backgroundColor: Colors.amber,
               color: Colors.red,
-              strokeWidth: 9.sp,
+              strokeWidth: 9,
             ),
           ),
           SizedBox(
-            height: 200.sp,
-            width: 200.sp,
-            child:
-                const CurvedLinearProgressIndicator(), //Has some error for now. TODO: fix this issue.
+            height: 200,
+            width: 200,
+            child: CurvedLinearProgressIndicator(),
           ),
           SizedBox(
-            height: 200.sp,
-            width: 200.sp,
-            child: const CurvedLinearProgressIndicator(
+            height: 200,
+            width: 200,
+            child: CurvedLinearProgressIndicator(
               value: 0.5,
+              strokeWidth: 8,
               backgroundColor: Colors.amber,
               color: Colors.red,
             ),