+++ CHANGELOG.md
@@ -1,3 +1,11 @@
-## 0.0.1
+## 1.0.0
 
-* TODO: Describe initial release.
+* Default flutter Progressbar with Curved Edges.
+
+## 1.0.0+1
+
+* Minor bug fixes and updated the documentations.
+
+## 1.0.1
+
+* CurvedLinearProgressIndicator Progress indicator misbehave. #1, issue fix.